Collapse ConfigPredicates namespace
DRY Validations does not appear to support configuring a schema with
multiple custom predicate modules.
